+++ CHANGELOG.md
@@ -2,15 +2,24 @@ Change Log
 ==========
 
 
+Version 2.1.0 *(2014-06-26)*
+----------------------------
+
+* Support Roboto Typefaces from Android L (Material Design).
+* Implementation of Alternative Attributes to Set the Typeface.
+* Use Sub-Pixel Rendering.
+* Support Gradle 0.11
+
+
 Version 2.0.1 *(2014-04-28)*
---------------------------
+----------------------------
 
 * Make typeface constants public.
 * Default to Roboto Regular instead of Thin when no typeface is specified.
 
 
 Version 2.0.0 *(2013-12-05)*
---------------------------
+----------------------------
 
 * Separated sample from library project.
 * Pushed to Maven Central.